refactor: remove unused session port interface (#62)
